lapcounter-server-test-mqtt-latency/requirements.txt
aiomqtt==1.2.1      # breaking changes in 2.0.0
paho-mqtt==1.6.1
python-dotenv==1.0.1
